moved beforeall to separate file
