+++ src/__tests__/__setup__/setup.js
@@ -1,5 +1,6 @@
 // import a variety of packages
 import { TextEncoder, TextDecoder } from "util";
 
+
 // Globalize TextDecoder & TextEncoder
 Object.assign(global, { TextDecoder, TextEncoder });
