+++ src/__tests__/pages/household/input/TaxYear.test.js
@@ -5,6 +5,7 @@ import fetch from "node-fetch";
 import TaxYear from "pages/household/input/TaxYear";
 import { defaultYear } from "data/constants";
 import { defaultHouseholds } from "data/defaultHouseholds";
+import { metadataUS } from "../../../__setup__/jest.test";
 
 jest.mock("react-router-dom", () => {
   const originalModule = jest.requireActual("react-router-dom");
@@ -21,13 +22,13 @@ useSearchParams.mockImplementation(() => {
 });
 
 describe("Test TaxYear component", () => {
-  let metadataUS = null;
+  // let metadataUS = null;
 
-  beforeAll(async () => {
-    const res = await fetch("https://api.policyengine.org/us/metadata");
-    const metadataRaw = await res.json();
-    metadataUS = metadataRaw.result;
-  });
+  // beforeAll(async () => {
+  //   const res = await fetch("https://api.policyengine.org/us/metadata");
+  //   const metadataRaw = await res.json();
+  //   metadataUS = metadataRaw.result;
+  // });
 
   test("Properly sets display years based on metadata", async () => {
     const setYear = jest.fn();
